+++ src/app/(home)/(profile)/_components/setting/profilemodal/CheckCurrentPassword.tsx
@@ -1,8 +1,9 @@
-import React, { useEffect, useState } from 'react';
+import { useState, useEffect } from 'react';
 import ReverseExclamation from '@/assets/images/common/ReverseExclamation';
 import X from '@/assets/images/common/X';
-import { useAuth } from '@/context/auth.context';
 import CircleX from '@/assets/images/common/CircleX';
+import { useAuth } from '@/context/auth.context';
+import useCapsLock from '@/hooks/common/useCapsLock';
 
 type CheckCurrentPasswordProps = {
   onValidationChange: (message: string) => void;
@@ -11,15 +12,13 @@ type CheckCurrentPasswordProps = {
 const CheckCurrentPassword = ({ onValidationChange }: CheckCurrentPasswordProps) => {
   const [currentPassword, setCurrentPassword] = useState<string>('');
   const [isCurrentPasswordValid, setIsCurrentPasswordValid] = useState<boolean>(true);
-  const [isValidating, setIsValidating] = useState<boolean>(false);
   const [validationMessage, setValidationMessage] = useState<string>('');
-  const [isCapsLockOn, setIsCapsLockOn] = useState<boolean>(false);
+  const [isFocused, setIsFocused] = useState<boolean>(false);
+
   const { me } = useAuth();
+  const isCapsLockOn = useCapsLock(isFocused);
 
-  // 현재 비밀번호 확인 요청
   const handleValidatePassword = async (password: string) => {
-    setIsValidating(true);
-
     try {
       const response = await fetch('/api/profile/verify-password', {
         method: 'POST',
@@ -35,18 +34,14 @@ const CheckCurrentPassword = ({ onValidationChange }: CheckCurrentPasswordProps)
         setValidationMessage('기존 비밀번호가 확인되었습니다.');
       } else {
         setIsCurrentPasswordValid(false);
-        setValidationMessage('기존 비밀번호가 올바르지 않습니다.');
+        setValidationMessage('비밀번호가 일치하지 않습니다.');
       }
-    } catch (error) {
-      console.error('error:', error);
+    } catch {
       setIsCurrentPasswordValid(false);
       setValidationMessage('서버 오류가 발생했습니다.');
-    } finally {
-      setIsValidating(false);
     }
   };
 
-  // 입력값이 변경될 때마다 유효성 검사 수행
   useEffect(() => {
     if (currentPassword.length > 0) {
       handleValidatePassword(currentPassword);
@@ -57,32 +52,25 @@ const CheckCurrentPassword = ({ onValidationChange }: CheckCurrentPasswordProps)
     onValidationChange(validationMessage);
   }, [currentPassword, validationMessage]);
 
-  useEffect(() => {
-    const handleKeyDown = (event: KeyboardEvent) => {
-      setIsCapsLockOn(event.getModifierState('CapsLock'));
-    };
-    const handleKeyUp = (event: KeyboardEvent) => {
-      setIsCapsLockOn(event.getModifierState('CapsLock'));
-    };
-    window.addEventListener('keydown', handleKeyDown);
-    window.addEventListener('keyup', handleKeyUp);
-    return () => {
-      window.removeEventListener('keydown', handleKeyDown);
-      window.removeEventListener('keyup', handleKeyUp);
-    };
-  }, []);
-
   return (
     <div className="h-[150px]">
       <p className="text-subtitle2 font-bold text-neutral-900 mb-1">기존 비밀번호</p>
       <div
-        className={`relative flex items-center justify-between w-[421px] h-[56px] p-[16px] border rounded mb-2 ${currentPassword.length === 0 ? 'text-neutral-700 outline-neutral-400 border border-neutral-300' : isCurrentPasswordValid ? 'text-neutral-900 outline-main-400 border border-main-400' : 'text-red outline-red border border-red'}`}
+        className={`relative flex items-center justify-between w-full max-w-lg h-14 p-4 border rounded mb-2 ${
+          currentPassword.length === 0
+            ? 'text-neutral-700 outline-neutral-400 border-neutral-300'
+            : isCurrentPasswordValid
+              ? 'text-neutral-900 outline-main-400 border-main-400'
+              : 'text-red outline-red border-red'
+        }`}
       >
         <input
           type="password"
           value={currentPassword}
           onChange={(e) => setCurrentPassword(e.target.value)}
-          className="w-[370px] outline-transparent"
+          onFocus={() => setIsFocused(true)}
+          onBlur={() => setIsFocused(false)}
+          className="w-full outline-transparent"
           placeholder="비밀번호를 입력해 주세요."
         />
         {currentPassword && (
@@ -92,15 +80,13 @@ const CheckCurrentPassword = ({ onValidationChange }: CheckCurrentPasswordProps)
         )}
       </div>
 
-      {isCapsLockOn ? (
+      {isCapsLockOn && (
         <div className="ml-1 my-2 flex items-center">
           <span>
             <ReverseExclamation stroke="#423edf" />
           </span>
           <span className="ml-1 text-body2 font-regular text-main-400">Caps Lock on</span>
         </div>
-      ) : (
-        ''
       )}
 
       {currentPassword.length > 3 && (
@@ -111,11 +97,6 @@ const CheckCurrentPassword = ({ onValidationChange }: CheckCurrentPasswordProps)
               {validationMessage}
             </span>
           </div>
-          {isValidating && (
-            <div className="flex items-center h-6 ">
-              <p className="text-body2 font-regular text-gray-700">검증 중...</p>
-            </div>
-          )}
         </>
       )}
     </div>
